# src/main/resources/openapi/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Kafka Data Aggregation API
  description: Real-time data aggregation system for multiple Kafka topics
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Real-Time Updates
    description: Server-Sent Events for real-time notifications
  - name: Sensor Data
    description: CRUD operations for sensor data

paths:
  /updates:
    get:
      tags:
        - Real-Time Updates
      summary: Subscribe to real-time topic updates
      description: |
        Server-Sent Events (SSE) endpoint that streams real-time notifications 
        whenever data is received from Kafka topics.
        
        Each event message format: "Hi, <TopicName> is sending the data from <Location>"
      operationId: streamUpdates
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
                example: "Hi, sensor-data-1 is sending the data from Location1"
              examples:
                sensorUpdate:
                  value: "Hi, sensor-data-1 is sending the data from Building-A"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/sensor-data:
    get:
      tags:
        - Sensor Data
      summary: Get all sensor data
      description: Retrieve all aggregated sensor data from the database
      operationId: getAllSensorData
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successfully retrieved sensor data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorData'

  /api/sensor-data/{id}:
    get:
      tags:
        - Sensor Data
      summary: Get sensor data by ID
      operationId: getSensorDataById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Sensor data found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorData'
        '404':
          description: Sensor data not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/sensor-data/topic/{topicName}:
    get:
      tags:
        - Sensor Data
      summary: Get sensor data by topic name
      operationId: getSensorDataByTopic
      parameters:
        - name: topicName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved sensor data for topic
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorData'

components:
  schemas:
    SensorData:
      type: object
      required:
        - topicName
        - location
        - data
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier
          example: 1
        topicName:
          type: string
          description: Kafka topic name
          example: sensor-data-1
        location:
          type: string
          description: Device location
          example: Building-A
        data:
          type: string
          description: JSON payload from the device
          example: '{"temperature": 25.5, "humidity": 60}'
        timestamp:
          type: string
          format: date-time
          description: Timestamp when data was received
          example: '2024-12-16T10:30:00Z'
        createdAt:
          type: string
          format: date-time
          description: Record creation time
        updatedAt:
          type: string
          format: date-time
          description: Record update time

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string